🩹 fix: 레스토랑 수정시 이미지 첨부 새로 안해도 수정 되도록

+++ src/main/java/com/sparta/goodbite/domain/restaurant/service/RestaurantService.java
@@ -72,13 +72,22 @@ public class RestaurantService {
         validateRestaurantOwnership(owner, restaurant);
 
         String originalImage = restaurant.getImageUrl();
-        String restaurantImage = s3Service.upload(image);
+        String restaurantImage = originalImage;
         try {
-            restaurant.update(restaurantRequestDto, restaurantImage);
-            s3Service.deleteImageFromS3(originalImage);
+            if (image != null) {
+                restaurantImage = s3Service.upload(image);
+
+                restaurant.update(restaurantRequestDto, restaurantImage);
+                s3Service.deleteImageFromS3(originalImage);
+            } else {
+                restaurant.update(restaurantRequestDto, originalImage);
+            }
         } catch (Exception e) {
-            s3Service.deleteImageFromS3(restaurantImage);
-            throw new RestaurantUpdateFailedException(RestaurantErrorCode.RESTAURANT_UPDATE_FAILED);
+            if (!restaurantImage.equals(originalImage)) {
+                s3Service.deleteImageFromS3(restaurantImage);
+            }
+            throw new RestaurantUpdateFailedException(
+                RestaurantErrorCode.RESTAURANT_UPDATE_FAILED);
         }
     }
